create implicit Join ID field with "bigint" type

+++ src/Persistence/Sql/Join.php
@@ -33,11 +33,11 @@ class Join extends Model\Join
         // TODO this mutates the owner model/joins!
         if (!$this->reverse && !$this->getOwner()->hasField($this->masterField)) {
             $owner = $this->hasJoin() ? $this->getJoin() : $this->getOwner();
-            $field = $owner->addField($this->masterField, ['type' => 'integer', 'system' => true, 'readOnly' => true]);
+            $field = $owner->addField($this->masterField, ['type' => 'bigint', 'system' => true, 'readOnly' => true]);
             $this->masterField = $field->shortName; // TODO this mutates the join!
         } elseif ($this->reverse && !$this->getOwner()->hasField($this->foreignField) && $this->hasJoin()) {
             $owner = $this->getJoin();
-            $field = $owner->addField($this->foreignField, ['type' => 'integer', 'system' => true, 'readOnly' => true, 'actual' => $this->masterField]);
+            $field = $owner->addField($this->foreignField, ['type' => 'bigint', 'system' => true, 'readOnly' => true, 'actual' => $this->masterField]);
             $this->foreignField = $field->shortName; // TODO this mutates the join!
         }
     }
